nanopore/promethion: drop 'pool' and add explicit 'runs' to model.
Drops the 'pool' attribute from all classes that previously referenced it, and adds a new class 'RunDir' to represent batches of samples run at the same time on the instrument. 'Pool' is also dropped from '__repr__' when reporting names.

'RunDir' now gathers the flowcells and basecalls dirs, with the 'ProjectDir' class updated to store multiple runs; the 'ProjectDir' interface is preserved by adding new properties 'flow_cells' and 'basecalls_dirs', which replicate the old attributes of the same name.