analysis: reimplement structure of analysis dirs in 'ProjectAnalysisDir'.
Reimplements the structure of PromethION analysis directories which are created and read by the 'ProjectAnalysisDir' class, to give each run its own number subdirectory with just the metadata etc for that run.

Removes the option to specific samples when creating the analysis directory; now placeholder 'samples.tsv' files are made for each run but have to be populated manually.