analysis: reimplement reporting from 'ProjectAnalysisDir' class.
Overhauls the 'report' method of the 'ProjectAnalysisDir' class, by making it a wrapper to one of two new methods 'report_project_summary' and 'report_runs'.

These changes remove the 'format' option and update the output in 'summary' mode to include details of the runs (it also drops the 'fields' option).

The 'runs' mode essential replaces the previous 'tsv' format, but now produces one line per run ('fields' must still be specified in this mode).