analysis: fix bug with reporting projects with no samples defined.
Fixes a bug in summary reporting by the 'report_project_summary' method, for projects and runs where there are no samples. Without the fix the summary reported that the runs contained one sample.

The fix also required an update to the reporting of sample names from the 'get_value' method of 'ProjectAnalysisDir', to return an empty string rather than '?' when there are no samples.