analysis: add 'update' method to the ProjectAnalysisDir class.
Adds a new method 'update' which rescans a PromethION data directory and updates the analysis directory with any runs that weren't present when the initial 'create' method was called.

As part of this update there is a new method 'import_runs', which pulls the code from 'create' for adding runs and makes it available to 'update'.